Klocwork_LS1021_ASF: Fixing klocwork issues.
This patch fixes the klocwork issues present in
ASF_SDK_LS1021A_V1-1.

CR: ENGR00331753
Reviewed-on: http://git.am.freescale.net:8181/19130
Tested-by: Hemant Agrawal <hemant@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/22647
Tested-by: Sandeep Malik <Sandeep.Malik@freescale.com>

+++ asfmodule/asfffp/driver/asftmr.h
@@ -26,7 +26,7 @@ struct asfTmr_s {
 	struct asfTmr_s *pNext;
 	struct asfTmr_s *pPrev;
 	unsigned long ulState;
-	unsigned short int ulCoreId;
+	unsigned int ulCoreId;
 	unsigned int ulBucketIndex;
 	unsigned int ulPoolId;
 	unsigned int ulTmOutVal;
